[main] автотесты

+++ controllers/cards.js
@@ -1,3 +1,4 @@
+const { default: mongoose } = require('mongoose');
 const Card = require('../models/card');
 
 const INPUT_ERROR = 400;
@@ -15,17 +16,19 @@ const createCard = (req, res) => {
   if (name && link) {
     Card.create({ name, link, owner: req.user._id })
       .then((card) => res.send(card))
-      .catch(() => res.status(INPUT_ERROR).send({ message: 'Неверно введены данные' }));
-  } else res.status(NOT_FOUND_ERROR).send({ message: 'Карточка не найдена' });
+      .catch(() => res.status(NOT_FOUND_ERROR).send({ message: 'Карточка не найдена' }));
+  } else res.status(INPUT_ERROR).send({ message: 'Неверно введены данные' });
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(() => {
-      res.status(INPUT_ERROR).send({ message: 'id карточки введен неверно' });
-    })
-    .then((card) => res.send({ data: card }))
-    .catch((error) => res.status(NOT_FOUND_ERROR).send({ message: 'Карточка не найдена' }, error));
+  if (mongoose.Types.ObjectId.isValid(req.params.cardId)) {
+    Card.findByIdAndRemove(req.params.cardId)
+      .orFail(() => {
+        res.status(INPUT_ERROR).send({ message: 'id карточки введен неверно' });
+      })
+      .then((card) => res.send({ data: card }))
+      .catch((error) => res.status(NOT_FOUND_ERROR).send({ message: 'Карточка не найдена' }, error));
+  } else res.status(INPUT_ERROR).send({ message: 'id карточки введен неверно' });
 };
 
 const likeCard = (req, res) => {
